Correctified the way of using frequencies and normalizeation factor for all V4 separable filters. The period of x, y ,u, v and r is assumed to be 2*sigma. The priod of theta is 2*pi. All freqencies are interge now.

+++ CV/V4Filter.hs
@@ -58,11 +58,11 @@ data V4QuadTreeSeparableFilterParams = V4QuadTreeSeparableFilterParams
   , polarSeparableAngularFreq    :: ![Int]
   , polarSeparableName           :: PolarSeparableFilterName
   , cartesianSeparableScale      :: ![Double]
-  , cartesianSeparableXFreq      :: ![Double]
-  , cartesianSeparableYFreq      :: ![Double]
+  , cartesianSeparableXFreq      :: ![Int]
+  , cartesianSeparableYFreq      :: ![Int]
   , hyperbolicSeparableScale     :: ![Double]
-  , hyperbolicSeparableUFreq     :: ![Double]
-  , hyperbolicSeparableVFreq     :: ![Double]
+  , hyperbolicSeparableUFreq     :: ![Int]
+  , hyperbolicSeparableVFreq     :: ![Int]
   , hyperbolicSeparableAngle     :: !Double
   } deriving (Show, Read)
